handle input events in engine on Android
