+++ cocos/platform/android/Android.mk
@@ -10,9 +10,10 @@ LOCAL_SRC_FILES := \
 CCDevice-android.cpp \
 CCFileUtils-android.cpp \
 CCApplication-android.cpp \
+CCView.cpp \
 CCCanvasRenderingContext2D-android.cpp \
 jni/JniImp.cpp \
-jni/JniHelper.cpp \
+jni/JniHelper.cpp
 
 LOCAL_EXPORT_C_INCLUDES := $(LOCAL_PATH)
