+++ templates/js-template-link/frameworks/runtime-src/proj.android-studio/app/jni/hellojavascript/main.cpp
@@ -27,9 +27,8 @@
 #include <android/log.h>
 #include <unordered_map>
 #include "platform/android/CCFileUtils-android.h"
-#include "platform/android/jni/JniHelper.h"
+#include "platform/android/CCView.h"
 #include "Game.h"
-#include "scripting/js-bindings/event/EventDispatcher.h"
 
 #define LOG_TAG "main"
 #define LOGD(...) __android_log_print(ANDROID_LOG_DEBUG, LOG_TAG, __VA_ARGS__)
@@ -37,167 +36,63 @@
 namespace
 {
     Game *game = nullptr;
-    struct cocos2d::TouchEvent touchEvent;
-    struct cocos2d::KeyboardEvent keyboardEvent;
 
     /**
-     * Shared state for our app.
-     */
+ * Shared state for our app.
+ */
     struct SavedState {
         struct android_app* app = nullptr;
 
         int animating = 0;
-        int32_t width = 0;
-        int32_t height = 0;
         bool initialized = false;
     };
+}
 
-    void engineHandleCmd(struct android_app* app, int32_t cmd)
+void handle_cmd(struct android_app* app, int32_t cmd)
+{
+    struct SavedState* state = (struct SavedState*)app->userData;
+    switch (cmd)
     {
-        struct SavedState* state = (struct SavedState*)app->userData;
-        switch (cmd)
-        {
-            case APP_CMD_INIT_WINDOW:
-                if (state->app->window)
-                {
-                    state->width = ANativeWindow_getWidth(app->window);
-                    state->height = ANativeWindow_getHeight(app->window);
-                    game = new Game(state->width, state->height);
-                    game->init();
-                    state->initialized = true;
-                }
-                break;
-            case APP_CMD_TERM_WINDOW:
-                state->animating = 0;
-                break;
-            case APP_CMD_GAINED_FOCUS:
-                state->animating = 1;
-                break;
-            case APP_CMD_LOST_FOCUS:
-                state->animating = 0;
-                break;
-            case APP_CMD_PAUSE:
-                if (state->initialized)
-                    cocos2d::Application::getInstance()->onPause();
-                break;
-            case APP_CMD_RESUME:
-                if (state->initialized)
-                    cocos2d::Application::getInstance()->onResume();
-                break;
-            default:
-                break;
-        }
-    }
-
-    void dispatchTouchEvent(int index, AInputEvent* event, cocos2d::TouchEvent& touchEvent) {
-        int pointerID = AMotionEvent_getPointerId(event, index);
-        touchEvent.touches.push_back({
-            AMotionEvent_getX(event, index), // x
-            AMotionEvent_getY(event, index), // y
-            pointerID});
-
-        cocos2d::EventDispatcher::dispatchTouchEvent(touchEvent);
-    }
-
-    void dispatchTouchEvents(AInputEvent* event, cocos2d::TouchEvent& touchEvent) {
-        size_t pointerCount = AMotionEvent_getPointerCount(event);
-        for (size_t i = 0; i < pointerCount; ++i) {
-            touchEvent.touches.push_back({
-                AMotionEvent_getX(event, i),
-                AMotionEvent_getY(event, i),
-                AMotionEvent_getPointerId(event, i)
-            });
-        }
-
-        cocos2d::EventDispatcher::dispatchTouchEvent(touchEvent);
-        touchEvent.touches.clear();
-    }
-
-    // key values in web, refer to http://docs.cocos.com/creator/api/en/enums/KEY.html
-    std::unordered_map<int, int> keyCodeMap = {
-            {AKEYCODE_BACK, 6},
-            {AKEYCODE_ENTER, 13},
-            {AKEYCODE_MENU, 18},
-            {AKEYCODE_DPAD_UP, 1003},
-            {AKEYCODE_DPAD_DOWN, 1004},
-            {AKEYCODE_DPAD_LEFT, 1000},
-            {AKEYCODE_DPAD_RIGHT, 1001},
-            {AKEYCODE_DPAD_CENTER, 1005}
-    };
-
-    /**
-     * Process the next input event.
-     */
-    int32_t engineHandleInput(struct android_app* app, AInputEvent* event) {
-        int type = AInputEvent_getType(event);
-
-        if(type == AINPUT_EVENT_TYPE_KEY){
-            int action = AKeyEvent_getAction(event);
-            if (action == AKEY_EVENT_ACTION_MULTIPLE)
-                return 0;
-
-            int keyCode = AKeyEvent_getKeyCode(event);
-            if (keyCodeMap.count(keyCode) > 0)
-                keyCode = keyCodeMap[keyCode];
-            else
-                keyCode = 0;
-
-            keyboardEvent.key = keyCode;
-            keyboardEvent.action = action == AKEY_EVENT_ACTION_DOWN
-                                             ? cocos2d::KeyboardEvent::Action::RELEASE
-                                             : cocos2d::KeyboardEvent::Action::PRESS;
-            cocos2d::EventDispatcher::dispatchKeyboardEvent(keyboardEvent);
-
-            return 1;
-        } else if (type == AINPUT_EVENT_TYPE_MOTION){
-            int action = AMotionEvent_getAction(event);
-
-            switch (action & AMOTION_EVENT_ACTION_MASK) {
-                case AMOTION_EVENT_ACTION_DOWN:
-                    touchEvent.type = cocos2d::TouchEvent::Type::BEGAN;
-                    dispatchTouchEvent(0, event, touchEvent);
-                    break;
-                case AMOTION_EVENT_ACTION_POINTER_DOWN:
-                    touchEvent.type = cocos2d::TouchEvent::Type::BEGAN;
-                    dispatchTouchEvent((action & AMOTION_EVENT_ACTION_POINTER_INDEX_MASK) >> AMOTION_EVENT_ACTION_POINTER_INDEX_SHIFT,
-                                        event,
-                                        touchEvent);
-                    break;
-                case AMOTION_EVENT_ACTION_UP:
-                    touchEvent.type = cocos2d::TouchEvent::Type::ENDED;
-                    dispatchTouchEvent(0, event, touchEvent);
-                    break;
-                case AMOTION_EVENT_ACTION_POINTER_UP:
-                    touchEvent.type = cocos2d::TouchEvent::Type::ENDED;
-                    dispatchTouchEvent((action & AMOTION_EVENT_ACTION_POINTER_INDEX_MASK) >> AMOTION_EVENT_ACTION_POINTER_INDEX_SHIFT,
-                                       event,
-                                       touchEvent);
-                    break;
-                case AMOTION_EVENT_ACTION_MOVE:
-                    touchEvent.type = cocos2d::TouchEvent::Type::MOVED;
-                    dispatchTouchEvents(event, touchEvent);
-                    break;
-                case AMOTION_EVENT_ACTION_CANCEL:
-                    touchEvent.type = cocos2d::TouchEvent::Type::CANCELLED;
-                    dispatchTouchEvents(event, touchEvent);
-                    break;
-                default:
-                    return 0;
+        case APP_CMD_INIT_WINDOW:
+            if (state->app->window)
+            {
+                auto width = ANativeWindow_getWidth(app->window);
+                auto height = ANativeWindow_getHeight(app->window);
+                game = new Game(width, height);
+                game->init();
+                state->initialized = true;
             }
-
-            return 1;
-        }
-
-        return 0;
+            break;
+        case APP_CMD_TERM_WINDOW:
+            state->animating = false;
+            break;
+        case APP_CMD_GAINED_FOCUS:
+            state->animating = true;
+            break;
+        case APP_CMD_LOST_FOCUS:
+            state->animating = false;
+            break;
+        case APP_CMD_PAUSE:
+            if (state->initialized)
+                cocos2d::Application::getInstance()->onPause();
+            break;
+        case APP_CMD_RESUME:
+            if (state->initialized)
+                cocos2d::Application::getInstance()->onResume();
+            break;
+        default:
+            break;
     }
+
+    cocos2d::View::engineHandleCmd(app, cmd);
 }
 
 void android_main(struct android_app* state) {
     struct SavedState savedState;
     memset(&savedState, 0, sizeof(savedState));
     state->userData = &savedState;
-    state->onAppCmd = engineHandleCmd;
-    state->onInputEvent = engineHandleInput;
+    state->onAppCmd = handle_cmd;
+    state->onInputEvent = cocos2d::View::engineHandleInput;
     savedState.app = state;
 
     cocos2d::JniHelper::setAndroidApp(state);
@@ -221,13 +116,9 @@ void android_main(struct android_app* state) {
         }
 
         if (state->destroyRequested != 0)
-        {
             return;
-        }
 
         if (savedState.animating)
-        {
             game->tick();
-        }
     }
 }
